+++ pdfviewer/src/iosMain/kotlin/tech/kotlinlang/pdfviewer/PdfViewer.ios.kt
@@ -1,0 +1,7 @@
+package tech.kotlinlang.pdfviewer
+
+import androidx.compose.ui.Modifier
+
+actual fun PdfViewerLoaded(modifier: Modifier, bytes: ByteArray) {
+    TODO("Not yet implemented")
+}